convert to ES6 and code clean up
